+++ flickit-assessment-data/src/main/java/org/flickit/assessment/data/jpa/core/evidence/EvidenceDescriptionAttachmentsCountView.java
@@ -4,5 +4,5 @@ public interface EvidenceDescriptionAttachmentsCountView {
 
     String getDescription();
 
-    Long getAttachmentsCount();
+    int getAttachmentsCount();
 }
